style: configure and run forge fmt

+++ src/StdStorage.sol
@@ -4,9 +4,8 @@ pragma solidity >=0.6.2 <0.9.0;
 import "./Vm.sol";
 
 struct StdStorage {
-    mapping (address => mapping(bytes4 => mapping(bytes32 => uint256))) slots;
-    mapping (address => mapping(bytes4 =>  mapping(bytes32 => bool))) finds;
-
+    mapping(address => mapping(bytes4 => mapping(bytes32 => uint256))) slots;
+    mapping(address => mapping(bytes4 => mapping(bytes32 => bool))) finds;
     bytes32[] _keys;
     bytes4 _sig;
     uint256 _depth;
@@ -15,18 +14,12 @@ struct StdStorage {
 }
 
 library stdStorageSafe {
-    event SlotFound(address who, bytes4 fsig, bytes32 keysHash, uint slot);
-    event WARNING_UninitedSlot(address who, uint slot);
+    event SlotFound(address who, bytes4 fsig, bytes32 keysHash, uint256 slot);
+    event WARNING_UninitedSlot(address who, uint256 slot);
 
     Vm private constant vm = Vm(address(uint160(uint256(keccak256("hevm cheat code")))));
 
-    function sigs(
-        string memory sigStr
-    )
-        internal
-        pure
-        returns (bytes4)
-    {
+    function sigs(string memory sigStr) internal pure returns (bytes4) {
         return bytes4(keccak256(bytes(sigStr)));
     }
 
@@ -36,12 +29,7 @@ library stdStorageSafe {
     //  if map, will be keccak256(abi.encode(key, uint(slot)));
     //  if deep map, will be keccak256(abi.encode(key1, keccak256(abi.encode(key0, uint(slot)))));
     //  if map struct, will be bytes32(uint256(keccak256(abi.encode(key1, keccak256(abi.encode(key0, uint(slot)))))) + structFieldDepth);
-    function find(
-        StdStorage storage self
-    )
-        internal
-        returns (uint256)
-    {
+    function find(StdStorage storage self) internal returns (uint256) {
         address who = self._target;
         bytes4 fsig = self._sig;
         uint256 field_depth = self._depth;
@@ -56,17 +44,20 @@ library stdStorageSafe {
         bytes32 fdat;
         {
             (, bytes memory rdat) = who.staticcall(cald);
-            fdat = bytesToBytes32(rdat, 32*field_depth);
+            fdat = bytesToBytes32(rdat, 32 * field_depth);
         }
 
-        (bytes32[] memory reads, ) = vm.accesses(address(who));
+        (bytes32[] memory reads,) = vm.accesses(address(who));
         if (reads.length == 1) {
             bytes32 curr = vm.load(who, reads[0]);
             if (curr == bytes32(0)) {
                 emit WARNING_UninitedSlot(who, uint256(reads[0]));
             }
             if (fdat != curr) {
-                require(false, "stdStorage find(StdStorage): Packed slot. This would cause dangerous overwriting and currently isn't supported.");
+                require(
+                    false,
+                    "stdStorage find(StdStorage): Packed slot. This would cause dangerous overwriting and currently isn't supported."
+                );
             }
             emit SlotFound(who, fsig, keccak256(abi.encodePacked(ins, field_depth)), uint256(reads[0]));
             self.slots[who][fsig][keccak256(abi.encodePacked(ins, field_depth))] = uint256(reads[0]);
@@ -83,7 +74,7 @@ library stdStorageSafe {
                 bytes memory rdat;
                 {
                     (success, rdat) = who.staticcall(cald);
-                    fdat = bytesToBytes32(rdat, 32*field_depth);
+                    fdat = bytesToBytes32(rdat, 32 * field_depth);
                 }
 
                 if (success && fdat == bytes32(hex"1337")) {
@@ -100,7 +91,10 @@ library stdStorageSafe {
             require(false, "stdStorage find(StdStorage): No storage use detected for target.");
         }
 
-        require(self.finds[who][fsig][keccak256(abi.encodePacked(ins, field_depth))], "stdStorage find(StdStorage): Slot(s) not found.");
+        require(
+            self.finds[who][fsig][keccak256(abi.encodePacked(ins, field_depth))],
+            "stdStorage find(StdStorage): Slot(s) not found."
+        );
 
         delete self._target;
         delete self._sig;
@@ -134,6 +128,7 @@ library stdStorageSafe {
         self._keys.push(bytes32(amt));
         return self;
     }
+
     function with_key(StdStorage storage self, bytes32 key) internal returns (StdStorage storage) {
         self._keys.push(key);
         return self;
@@ -154,7 +149,6 @@ library stdStorageSafe {
         return abi.decode(read(self), (bytes32));
     }
 
-
     function read_bool(StdStorage storage self) internal returns (bool) {
         int256 v = read_int(self);
         if (v == 0) return false;
@@ -174,18 +168,17 @@ library stdStorageSafe {
         return abi.decode(read(self), (int256));
     }
 
-    function bytesToBytes32(bytes memory b, uint offset) private pure returns (bytes32) {
+    function bytesToBytes32(bytes memory b, uint256 offset) private pure returns (bytes32) {
         bytes32 out;
 
         uint256 max = b.length > 32 ? 32 : b.length;
-        for (uint i = 0; i < max; i++) {
+        for (uint256 i = 0; i < max; i++) {
             out |= bytes32(b[offset + i] & 0xFF) >> (i * 8);
         }
         return out;
     }
 
-    function flatten(bytes32[] memory b) private pure returns (bytes memory)
-    {
+    function flatten(bytes32[] memory b) private pure returns (bytes memory) {
         bytes memory result = new bytes(b.length * 32);
         for (uint256 i = 0; i < b.length; i++) {
             bytes32 k = b[i];
@@ -202,22 +195,11 @@ library stdStorageSafe {
 library stdStorage {
     Vm private constant vm = Vm(address(uint160(uint256(keccak256("hevm cheat code")))));
 
-    function sigs(
-        string memory sigStr
-    )
-        internal
-        pure
-        returns (bytes4)
-    {
+    function sigs(string memory sigStr) internal pure returns (bytes4) {
         return stdStorageSafe.sigs(sigStr);
     }
 
-    function find(
-        StdStorage storage self
-    )
-        internal
-        returns (uint256)
-    {
+    function find(StdStorage storage self) internal returns (uint256) {
         return stdStorageSafe.find(self);
     }
 
@@ -240,6 +222,7 @@ library stdStorage {
     function with_key(StdStorage storage self, uint256 amt) internal returns (StdStorage storage) {
         return stdStorageSafe.with_key(self, amt);
     }
+
     function with_key(StdStorage storage self, bytes32 key) internal returns (StdStorage storage) {
         return stdStorageSafe.with_key(self, key);
     }
@@ -265,10 +248,7 @@ library stdStorage {
         checked_write(self, t);
     }
 
-    function checked_write(
-        StdStorage storage self,
-        bytes32 set
-    ) internal {
+    function checked_write(StdStorage storage self, bytes32 set) internal {
         address who = self._target;
         bytes4 fsig = self._sig;
         uint256 field_depth = self._depth;
@@ -283,12 +263,15 @@ library stdStorage {
         bytes32 fdat;
         {
             (, bytes memory rdat) = who.staticcall(cald);
-            fdat = bytesToBytes32(rdat, 32*field_depth);
+            fdat = bytesToBytes32(rdat, 32 * field_depth);
         }
         bytes32 curr = vm.load(who, slot);
 
         if (fdat != curr) {
-            require(false, "stdStorage find(StdStorage): Packed slot. This would cause dangerous overwriting and currently isn't supported.");
+            require(
+                false,
+                "stdStorage find(StdStorage): Packed slot. This would cause dangerous overwriting and currently isn't supported."
+            );
         }
         vm.store(who, slot, set);
         delete self._target;
@@ -318,19 +301,18 @@ library stdStorage {
     }
 
     // Private function so needs to be copied over
-    function bytesToBytes32(bytes memory b, uint offset) private pure returns (bytes32) {
+    function bytesToBytes32(bytes memory b, uint256 offset) private pure returns (bytes32) {
         bytes32 out;
 
         uint256 max = b.length > 32 ? 32 : b.length;
-        for (uint i = 0; i < max; i++) {
+        for (uint256 i = 0; i < max; i++) {
             out |= bytes32(b[offset + i] & 0xFF) >> (i * 8);
         }
         return out;
     }
 
     // Private function so needs to be copied over
-    function flatten(bytes32[] memory b) private pure returns (bytes memory)
-    {
+    function flatten(bytes32[] memory b) private pure returns (bytes memory) {
         bytes memory result = new bytes(b.length * 32);
         for (uint256 i = 0; i < b.length; i++) {
             bytes32 k = b[i];
